requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg[binary]==3.2.3
gunicorn==23.0.0
whitenoise==6.11.0
dj-database-url==3.0.1
sqlparse==0.5.5
PyJWT==2.10.1
djangorestframework-simplejwt==5.3.1
google-generativeai>=0.8.0
google-auth>=2.0.0
requests>=2.0.0
python-dotenv>=1.0.0
